fix issues in the ManageEvents , making it easier to schedule new session for participant
